% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cis_eff_filter.R
\name{cis_eff_filter}
\alias{cis_eff_filter}
\title{filtering cisSNP-effect}
\usage{
cis_eff_filter(x, y, calc.idx, Bsg.thr = 0)
}
\arguments{
\item{x}{an gentype pair data in list form}

\item{y}{an expression data in matrix form}

\item{calc.idx}{the result value obtained by using the cor_filter function}

\item{Bsg.thr}{an threshold to filter cis-effect that are too small, default is 0}
}
\description{
This function is used to filter out small values of cisSNP-effect.
}
\examples{
x <- yeast_SNP
y <- yeast_exp

result <- cor_filter(x = x, y = y, cor.thr = 0 )

cis_eff_filter(x = x, y = y, calc.idx=result, Bsg.thr = 0)

}
