% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_filter.R
\name{cor_filter}
\alias{cor_filter}
\title{filtering gene-gene correlations}
\usage{
cor_filter(x, y, cor.thr = 0)
}
\arguments{
\item{x}{an gentype pair data in list form}

\item{y}{an expression data in matrix form}

\item{cor.thr}{an threshold to filter gene-gene correlations that are too small, default is 0}
}
\description{
This function is used to filter out small values of gene-gene correlation.
}
\examples{
x <- yeast_SNP
y <- yeast_exp

cor_filter(x = x, y = y, cor.thr = 0 )

}
