% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{expression}
\alias{expression}
\alias{yeast_exp}
\title{Example gene expression data set}
\format{
A data frame with 1012 rows and 82 columns
}
\source{
<https://figshare.com/s/83bddc1ddf3f97108ad4>
}
\usage{
yeast_exp()
}
\description{
This dataset contains data for demonstaration.
}
