% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fine_mapping.R
\name{FineMapping}
\alias{FineMapping}
\title{FineMapping function}
\usage{
FineMapping(x, y, pip.thr = 0)
}
\arguments{
\item{x}{an genotype data in double matrix}

\item{y}{an expression data in duble matrix}

\item{pip.thr}{an PIP threshold}
}
\description{
This function identifies causal variants to be used as instrumental variables.
}
\examples{

library("dplyr")
set.seed(1000)
x = sapply(1:10, function(i){
  rbinom(100, 2, 0.3) \%>\% as.double()
  }) ; colnames(x) = paste0("rs", 1:ncol(x))
y = matrix(0.5*scale(x)[,1] + 0.5*scale(x)[,2] + rnorm(100))
x_paris = FineMapping(x, y)
head(x_paris)

}
\keyword{fine}
\keyword{mapping}
