% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{gene_pair}
\alias{gene_pair}
\alias{yeast_SNP}
\title{Example gene-gene pair Dataset}
\format{
A data frame with 1012 rows and 780 columns
}
\source{
<https://figshare.com/s/83bddc1ddf3f97108ad4>
}
\usage{
yeast_SNP()
}
\description{
This dataset contains data for demonstaration.
}
