% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRN_cluster.R
\name{GRN_clustering_focus}
\alias{GRN_clustering_focus}
\title{GRN_clustering_focus function}
\usage{
GRN_clustering_focus(nodes, edges)
}
\arguments{
\item{nodes}{means genes}

\item{edges}{indicates gene-gene interactions}
}
\description{
It is a function that represents the interaction of genes in a cluster plot using a clustering algorithm.
}
\examples{
x <- yeast_SNP
y <- yeast_exp

result <- cor_filter(x, y)
result <- cis_eff_filter(x, y, result)
MRggi <- MGI(x, y, calc.idx = result,c.size = 5, GRN = FALSE)

grn <- GRN_clustering_focus(MRggi$nodes, MRggi$edges)

}
