% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGI.R
\name{MGI}
\alias{MGI}
\title{MGI function}
\usage{
MGI(
  x,
  y,
  calc.idx,
  p.adjust.method = "bonferroni",
  fdrth = 0.05,
  c.size = 10,
  GRN = F
)
}
\arguments{
\item{x}{an genotype pair in list form}

\item{y}{an expression data (n x k matrix, n = sample number, k = gene number)}

\item{calc.idx}{the data result value after correlation filtering and cis-effect filtering}

\item{p.adjust.method}{multiple correlation method}

\item{fdrth}{FDR threshold}

\item{c.size}{the number of genes constructing one cluster}

\item{GRN}{show clusters using clustering algorithm}
}
\description{
This function identifies gene-gene interactions for all gene-gene pairs that genes can have.
}
\examples{

x <- yeast_SNP
y <- yeast_exp

result <- cor_filter(x, y)
result <- cis_eff_filter(x, y, result)

MRggi <- MGI(x, y, calc.idx = result, fdrth = 0.03, c.size = 5, GRN = TRUE)

head(MRggi$data$edges) ; head(MRggi$data$nodes)
MRggi$GRN

}
