% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRggi-package.R, R/package.R
\docType{package}
\name{MRggi-package}
\alias{MRggi}
\alias{MRggi-package}
\title{MRggi: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.

More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
