% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multiple_regression.R
\name{Multiple_regression}
\alias{Multiple_regression}
\title{Multiple_regression Function}
\usage{
Multiple_regression(x, y)
}
\arguments{
\item{x}{an genotype pair data}

\item{y}{an expression data}
}
\description{
This function is used for regression analysis in MR calculations and requires one or more variants.
}
\examples{

library(dplyr)
set.seed(1000)
x = sapply(1:10, function(i){
  rbinom(100, 2, 0.3) \%>\% as.double()
  }) ; colnames(x) = paste0("rs", 1:ncol(x))
y = matrix(0.5*scale(x)[,1] + 0.5*scale(x)[,2] + rnorm(100))
x_paris = FineMapping(x, y)

beta_coef = Multiple_regression(x = x_paris$S1, y = y)

}
