% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quartet_2SLS.R
\name{Quartet_2SLS}
\alias{Quartet_2SLS}
\title{Quartet_2SLS Function}
\usage{
Quartet_2SLS(x1, x2, y1, y2)
}
\arguments{
\item{x1}{an genotype pair data}

\item{x2}{an genotype pair data}

\item{y1}{an expression data}

\item{y2}{an expression data}
}
\description{
This function is used for regression analysis in MR calculation and obtains p-values using the Wald test.
}
\examples{

library(dplyr)
set.seed(1000)
x = sapply(1:10, function(i){
  rbinom(100, 2, 0.3) \%>\% as.double()
  }) ; colnames(x) = paste0("rs", 1:ncol(x))
y = matrix(0.5*scale(x)[,1] + 0.5*scale(x)[,2] + rnorm(100))
x_paris = FineMapping(x, y)

coef <- Quartet_2SLS(x1 = x_paris$S1, x2 = x_paris$S1, y1 = y, y2 = y)

}
